+++ packages/core/src/utils/__tests__/get-variable.tests.ts
@@ -11,8 +11,13 @@ describe("getVariable", () => {
     expect(getVariable("org-package", "mode")).not.toBe(
       getVariable("org.package", "mode")
     );
-    expect(getVariable("@org/package", "html")).not.toBe(
+    expect(getVariable("@org/package", "default")).not.toBe(
       getVariable("@org/package", "amp")
     );
+  });
+  it("should allow scopes starting with numbers", () => {
+    expect(() =>
+      eval("var " + getVariable("@123/package", "mode"))
+    ).not.toThrow();
   });
 });